+++ addon/stores/base-store.js
@@ -445,6 +445,33 @@ export default DS.Store.extend({
   },
 
   /**
+    A method to get array of models with batch request.
+
+    @method batchQuery
+    @param {String} type Model name.
+    @param {Query} query Flexberry Query object.
+    @param {Boolean} [useOnlineStore] Allow to explicitly specify online or offline store using independently of global online status.
+    @return {Promise} A promise that fulfilled with an array of models.
+  */
+  batchQuery() {
+    return this._callSuperMethod('batchQuery', 2, arguments);
+  },
+
+  /**
+    A method to get single record with batch request.
+
+    @method batchFindRecord
+    @param {String} modelName Model name.
+    @param {String} modelId Record id.
+    @param {String} projectionName Projection name.
+    @param {Boolean} [useOnlineStore] Allow to explicitly specify online or offline store using independently of global online status.
+    @return {Promise} A promise that fulfilled with single record.
+  */
+  batchFindRecord() {
+    return this._callSuperMethod('batchFindRecord', 3, arguments);
+  },
+
+  /**
     Returns an instance of the adapter for a given type.
     @method adapterFor
     @param {String} modelName